+++ RaiderEngine/demos/demoParticleEmitter2Ds/RainEmitter.cpp
@@ -13,14 +13,14 @@ RainEmitter::RainEmitter(glm::vec2 pos) : ParticleEmitter2D(pos, "rain.png") {
 	spawnScaleMin = .75f, spawnScaleMax = 1;
 	spawnSpeedMin = 480, spawnSpeedMax = 600;
 	spawnAngleMin = glm::half_pi<float>(), spawnAngleMax = glm::half_pi<float>();
-	spawnMinLife = std::numeric_limits<double>::infinity(), spawnMaxLife = std::numeric_limits<double>::infinity();
+	spawnMinLife = std::numeric_limits<float>::infinity(), spawnMaxLife = std::numeric_limits<float>::infinity();
 	shrink = false;
 	fade = false;
 	circleSpawn = false;
 }
 
 void RainEmitter::update() {
-	for (int i = 0; i < particles.size(); ++i) {
+	for (unsigned int i = 0; i < particles.size(); ++i) {
 		if (particles[i].pos.y > SCR_HEIGHT + sprite.height / 2 && particleMotions[i].life > 0)
 			i-= killParticle(i);
 	}
